refs #296 build unit tests too
git-svn-id: https://devel.codership.com/svn/galera/trunk@1577 9148e567-b586-4176-b9b8-4e36cb4ec61b

+++ galerautils/tests/gu_fifo_test.c
@@ -11,8 +11,8 @@
 START_TEST (gu_fifo_test)
 {
     gu_fifo_t* fifo;
-    long i;
-    long* item;
+    size_t i;
+    size_t* item;
     size_t used;
 
     fifo = gu_fifo_create (0, 1);
